+++ backend/app/models/merchant_qualification.py
@@ -1,5 +1,5 @@
 from datetime import datetime, date
-from sqlalchemy import BigInteger, String, Date, Enum as SAEnum, ForeignKey, DateTime, func
+from sqlalchemy import Integer, String, Date, Enum as SAEnum, ForeignKey, DateTime, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from typing import Optional
 from app.database import Base
@@ -9,9 +9,9 @@ from app.models.product_qualification import QualStatus
 class MerchantQualification(Base):
     __tablename__ = "merchant_qualifications"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    merchant_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.id"), nullable=False)
-    qual_type_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("qualification_types.id"), nullable=False)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    merchant_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id"), nullable=False)
+    qual_type_id: Mapped[int] = mapped_column(Integer, ForeignKey("qualification_types.id"), nullable=False)
     file_url: Mapped[str] = mapped_column(String(500), nullable=False)
     file_name: Mapped[str] = mapped_column(String(255), default="")
     expire_date: Mapped[Optional[date]] = mapped_column(Date, nullable=True)
